fix perms check for GMemberAdd / Remove

+++ eventHandlers/guildMemberAdd.js
@@ -49,7 +49,9 @@ module.exports = async (guild, member) => {
     const url = `${paths.GENERATORS}/userio/in/${member.id}/${welcomeSkin || "minimal"}.png?text=${encodeURIComponent(txt)}`;
 
     resolveFile(url).then(async (buffer) => {
-      PLX.getChannel(welcomeChannel).send({ content: welcomeText, embed }, (welcomeImage ? file(buffer, "in.png") : null)).then((ms) => {
+      const welcomeChannelObj = PLX.getChannel(welcomeChannel);
+      if ( !welcomeChannelObj.permissionsOf(PLX.user.id).has('viewChannel') || !welcomeChannelObj.permissionsOf(PLX.user.id).has('sendMessages') ) return;
+      welcomeChannelObj.send({ content: welcomeText, embed }, (welcomeImage ? file(buffer, "in.png") : null)).then((ms) => {
         if (welcomeTimer) ms.deleteAfter(welcomeTimer).catch(() => null);
       }).catch(console.error);
     }).catch(console.error);
